feat: ensemble fragments in CreateEventActivity using pop up, renamed containers
