+++ app/src/main/java/com/example/calitour/activities/CreateEventActivity.kt
@@ -2,6 +2,7 @@ package com.example.calitour.activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import com.example.calitour.R
 import com.example.calitour.activities.fragments.CreateEventFragment
@@ -10,18 +11,27 @@ import com.example.calitour.databinding.ActivityCreateEventBinding
 
 class CreateEventActivity : AppCompatActivity() {
 
-    private val createEventFragment = CreateEventFragment.newInstance()
-    private val createProductFragment = CreateProductFragment.newInstance()
+    private val createEventFragment = CreateEventFragment()
+    private val createProductFragment = CreateProductFragment()
 
     private val binding: ActivityCreateEventBinding by lazy {
         ActivityCreateEventBinding.inflate(layoutInflater)
     }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        when(intent?.extras?.getString("fragment")){
+            "EVENT" ->{
+                showFragment(createEventFragment)
+            }
+            "PRODUCT" -> {
+                showFragment(createProductFragment)
+            }
+        }
         setContentView(binding.root)
     }
 
     fun showFragment(fragment: Fragment){
-        supportFragmentManager.beginTransaction().replace(R.id.fragmentContainer,fragment).commit()
+        supportFragmentManager.beginTransaction().replace(R.id.fragmentContainerCreate,fragment).commit()
     }
 }
