+++ app/src/main/res/layout/create_product_fragment.xml
@@ -88,7 +88,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:contentDescription="@string/name_of_your_event"
-                android:src="@tools:sample/avatars" />
+                />
         </androidx.cardview.widget.CardView>
 
         <Button
