--- app/src/main/res/values-es/strings.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="app_name">CaliTour</string>
-    <string name="contact_us">Contáctanos</string>
-    <string name="event">Evento</string>
-    <string name="product_sale">Producto en venta</string>
-    <string name="sign_in">Iniciar sesión</string>
-    <string name="dont_have_account">No tengo cuenta</string>
-    <string name="register_me">Registrarme</string>
-    <string name="name">Nombre</string>
-    <string name="email">E-mail</string>
-    <string name="password">Contraseña</string>
-    <string name="phone">Celular</string>
-    <string name="birthday">Fecha de nacimiento</string>
-    <string name="select_image">Selecciona una imagen</string>
-</resources>
